+++ infrastructures/ProjectBuilding/NpmHook.php
@@ -1,0 +1,93 @@
+<?php
+
+/*
+ * East Paas.
+ *
+ * LICENSE
+ *
+ * This source file is subject to the MIT license
+ * it is available in LICENSE file at the root of this package
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to richard@teknoo.software so we can send you a copy immediately.
+ *
+ *
+ * @copyright   Copyright (c) EIRL Richard Déloge (https://deloge.io - richard@deloge.io)
+ * @copyright   Copyright (c) SASU Teknoo Software (https://teknoo.software - contact@teknoo.software)
+ *
+ * @link        http://teknoo.software/east/paas Project website
+ *
+ * @license     http://teknoo.software/license/mit         MIT License
+ * @author      Richard Déloge <richard@teknoo.software>
+ */
+
+declare(strict_types=1);
+
+namespace Teknoo\East\Paas\Infrastructures\ProjectBuilding;
+
+use RuntimeException;
+use Symfony\Component\Process\Process;
+use Teknoo\East\Paas\Infrastructures\ProjectBuilding\Exception\InvalidArgumentException;
+use Teknoo\Recipe\Promise\PromiseInterface;
+use Teknoo\East\Paas\Contracts\Hook\HookInterface;
+use Throwable;
+
+use function array_merge;
+use function preg_match;
+use function reset;
+
+/**
+ * Hook to perform some operations with npm to install dependencies for Javascript Project.
+ * Available composer's commands are :
+ * - login
+ * - install
+ * - update
+ * - test
+ *
+ * @copyright   Copyright (c) EIRL Richard Déloge (https://deloge.io - richard@deloge.io)
+ * @copyright   Copyright (c) SASU Teknoo Software (https://teknoo.software - contact@teknoo.software)
+ * @license     http://teknoo.software/license/mit         MIT License
+ * @author      Richard Déloge <richard@teknoo.software>
+ */
+class NpmHook extends AbstractHook
+{
+    /**
+     * @param array{action?: string|null, arguments?: iterable<string>} $options
+     * @return string[]
+     * @throws InvalidArgumentException
+     */
+    protected function validateOptions(array $options): array
+    {
+
+        $grantedCommands = [
+            'login' => true,
+            'install' => true,
+            'update' => true,
+            'test' => true,
+        ];
+
+        $args = [];
+        if (!isset($options['action'])) {
+            $cmd = (string) reset($options);
+        } else {
+            $cmd = $options['action'];
+            $args = $options['arguments'] ?? [];
+        }
+
+        foreach ([$cmd, ...$args] as &$value) {
+            if (!is_scalar($value)) {
+                throw new InvalidArgumentException('composer action and arguments must be scalars values');
+            }
+
+            if (preg_match('#[\&\|<>;]#S', (string) $value)) {
+                throw new InvalidArgumentException('Pipe and redirection are forbidden');
+            }
+        }
+
+        if (!isset($grantedCommands[$cmd])) {
+            throw new InvalidArgumentException("$cmd is forbidden");
+        }
+
+        return [$cmd, ...$args];
+    }
+}