+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Teknoo Software - PaaS - Change Log
 
+## [1.8.0] - 2023-06-13
+### Stable Release
+- Convert Composer namespace to ProjectBuild to host any dependencies managers
+- Initial support of npm, pip and make tools
+- DI key `teknoo.east.paas.composer.phar.path` replaced by `teknoo.east.paas.composer.path`.
+  - `teknoo.east.paas.composer.phar.path` is deprecated.
+
 ## [1.7.1] - 2023-06-07
 ### Stable Release
 - Update Teknoo libs